SWE-Agent instance and configuration.
The agent was deeply modified in its configuration and bundle tools to adapt it for a more object oriented approach in the languages C, C+, Java. The main prompt and documentation was completely refactored, and we had to experiment with different temperatures and top_p values. We also created a couple of scripts to help both with the model server with ollama in modal and the json structure of the Pass@K calls for the agent, cleaning any dirty output (mainly bin files).

Co-authored-by: Juan José Salazar Cortéz <jjsc2003@live.com>